Fixed hits sg test
