+++ cpp/src/link_analysis/hits_sg.cu
@@ -19,76 +19,70 @@
 namespace cugraph {
 
 // SG instantiation
-template
-std::tuple<float, size_t>
-hits(raft::handle_t const& handle,
-     graph_view_t<int32_t, int32_t, float, true, false> const& graph_view,
-     float * const hubs,
-     float * const authorities,
-     float epsilon,
-     size_t max_iterations,
-     bool has_initial_hubs_guess,
-     bool normalize,
-     bool do_expensive_check);
+template std::tuple<float, size_t> hits(
+  raft::handle_t const& handle,
+  graph_view_t<int32_t, int32_t, float, true, false> const& graph_view,
+  float* const hubs,
+  float* const authorities,
+  float epsilon,
+  size_t max_iterations,
+  bool has_initial_hubs_guess,
+  bool normalize,
+  bool do_expensive_check);
 
-template
-std::tuple<double, size_t>
-hits(raft::handle_t const& handle,
-     graph_view_t<int32_t, int32_t, double, true, false> const& graph_view,
-     double * const hubs,
-     double * const authorities,
-     double epsilon,
-     size_t max_iterations,
-     bool has_initial_hubs_guess,
-     bool normalize,
-     bool do_expensive_check);
+template std::tuple<double, size_t> hits(
+  raft::handle_t const& handle,
+  graph_view_t<int32_t, int32_t, double, true, false> const& graph_view,
+  double* const hubs,
+  double* const authorities,
+  double epsilon,
+  size_t max_iterations,
+  bool has_initial_hubs_guess,
+  bool normalize,
+  bool do_expensive_check);
 
-template
-std::tuple<float, size_t>
-hits(raft::handle_t const& handle,
-     graph_view_t<int32_t, int64_t, float, true, false> const& graph_view,
-     float * const hubs,
-     float * const authorities,
-     float epsilon,
-     size_t max_iterations,
-     bool has_initial_hubs_guess,
-     bool normalize,
-     bool do_expensive_check);
+template std::tuple<float, size_t> hits(
+  raft::handle_t const& handle,
+  graph_view_t<int32_t, int64_t, float, true, false> const& graph_view,
+  float* const hubs,
+  float* const authorities,
+  float epsilon,
+  size_t max_iterations,
+  bool has_initial_hubs_guess,
+  bool normalize,
+  bool do_expensive_check);
 
-template
-std::tuple<double, size_t>
-hits(raft::handle_t const& handle,
-     graph_view_t<int32_t, int64_t, double, true, false> const& graph_view,
-     double * const hubs,
-     double * const authorities,
-     double epsilon,
-     size_t max_iterations,
-     bool has_initial_hubs_guess,
-     bool normalize,
-     bool do_expensive_check);
+template std::tuple<double, size_t> hits(
+  raft::handle_t const& handle,
+  graph_view_t<int32_t, int64_t, double, true, false> const& graph_view,
+  double* const hubs,
+  double* const authorities,
+  double epsilon,
+  size_t max_iterations,
+  bool has_initial_hubs_guess,
+  bool normalize,
+  bool do_expensive_check);
 
-template
-std::tuple<float, size_t>
-hits(raft::handle_t const& handle,
-     graph_view_t<int64_t, int64_t, float, true, false> const& graph_view,
-     float * const hubs,
-     float * const authorities,
-     float epsilon,
-     size_t max_iterations,
-     bool has_initial_hubs_guess,
-     bool normalize,
-     bool do_expensive_check);
+template std::tuple<float, size_t> hits(
+  raft::handle_t const& handle,
+  graph_view_t<int64_t, int64_t, float, true, false> const& graph_view,
+  float* const hubs,
+  float* const authorities,
+  float epsilon,
+  size_t max_iterations,
+  bool has_initial_hubs_guess,
+  bool normalize,
+  bool do_expensive_check);
 
-template
-std::tuple<double, size_t>
-hits(raft::handle_t const& handle,
-     graph_view_t<int64_t, int64_t, double, true, false> const& graph_view,
-     double * const hubs,
-     double * const authorities,
-     double epsilon,
-     size_t max_iterations,
-     bool has_initial_hubs_guess,
-     bool normalize,
-     bool do_expensive_check);
+template std::tuple<double, size_t> hits(
+  raft::handle_t const& handle,
+  graph_view_t<int64_t, int64_t, double, true, false> const& graph_view,
+  double* const hubs,
+  double* const authorities,
+  double epsilon,
+  size_t max_iterations,
+  bool has_initial_hubs_guess,
+  bool normalize,
+  bool do_expensive_check);
 
 }  // namespace cugraph
