+++ cpp/tests/CMakeLists.txt
@@ -383,7 +383,7 @@ ConfigureTest(SSSP_TEST traversal/sssp_test.cpp)
 
 ###################################################################################################
 # - HITS tests ------------------------------------------------------------------------------------
-ConfigureTest(HITS_TEST link_analysis/hits.cpp)
+ConfigureTest(HITS_TEST link_analysis/hits_test.cpp)
 
 ###################################################################################################
 # - PAGERANK tests --------------------------------------------------------------------------------
@@ -466,6 +466,10 @@ if(BUILD_CUGRAPH_MG_TESTS)
         ###########################################################################################
         # - MG PAGERANK tests ---------------------------------------------------------------------
         ConfigureTestMG(MG_PAGERANK_TEST link_analysis/mg_pagerank_test.cpp)
+
+        ###########################################################################################
+        # - MG HITS tests -------------------------------------------------------------------------
+        ConfigureTestMG(MG_HITS_TEST link_analysis/mg_hits_test.cpp)
 
         ###########################################################################################
         # - MG KATZ CENTRALITY tests --------------------------------------------------------------
